+++ src/Rocketeer/Rocketeer.php
@@ -26,7 +26,7 @@ class Rocketeer
      *
      * @type string
      */
-    const VERSION = '2.0.6';
+    const VERSION = '2.0.7';
 
     /**
      * The specific commit built.